pyright cleanup
